Cleanups in CBOR functionality (#676)
